Rename back to AuthConnectors
