+++ packages/teleport/src/AuthConnectors/ConnectorList/ConnectorList.tsx
@@ -19,7 +19,7 @@ import { Text, Flex, ButtonPrimary } from 'design';
 import * as Icons from 'design/Icon';
 import { MenuIcon, MenuItem } from 'shared/components/MenuAction';
 import { State as ResourceState } from 'teleport/components/useResources';
-import { State as GithubConnectorState } from '../useGithubConnectors';
+import { State as AuthConnectorState } from '../useAuthConnectors';
 
 export default function ConnectorList({ items, onEdit, onDelete }: Props) {
   items = items || [];
@@ -106,7 +106,7 @@ const menuActionProps = {
 };
 
 type Props = {
-  items: GithubConnectorState['items'];
+  items: AuthConnectorState['items'];
   onEdit: ResourceState['edit'];
   onDelete: ResourceState['remove'];
 };
