+++ packages/teleport/src/AuthConnectors/DeleteConnectorDialog/index.ts
