+++ packages/teleport/src/AuthConnectors/index.ts
@@ -14,5 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import GithubConnectors from './GithubConnectors';
-export default GithubConnectors;
+import AuthConnectors from './AuthConnectors';
+export default AuthConnectors;
